implemented popular

+++ server/routes/uploadRoutes.js
@@ -20,17 +20,8 @@ module.exports = (function () {
         let user = await req.user;
         let fileModel = await upload.findWithID(req.body.mongoID)
         if (req.files && req.files.thumbnail) {
-            /*let buffer = await req.files.thumbnail.data
-            sharp(buffer).toFile("./uploaded/thumbnails/" + fileModel._id + ".jpg");*/
-
-            console.log(resolve("uploaded/thumbnails/" + fileModel._id + "jpg"))
-
-            if (fs.existsSync(resolve("uploaded/thumbnails/" + fileModel._id + ".jpg"))) {
-                console.log("Removed file")
-                fs.unlinkSync(resolve("uploaded/thumbnails/" + fileModel._id + ".jpg"));
-            }
-
-            await req.files.thumbnail.mv("./uploaded/thumbnails/" + fileModel._id + extname(req.files.thumbnail.name)); //Moves file from tmp to server
+            if (fs.existsSync(resolve("uploaded/thumbnails/" + fileModel._id + ".jpg"))) fs.unlinkSync(resolve("uploaded/thumbnails/" + fileModel._id + ".jpg"));
+            await req.files.thumbnail.mv("./uploaded/thumbnails/" + fileModel._id + extname(req.files.thumbnail.name));
         }
         let thumbnailName = req.files && req.files.thumbnail ? req.files.thumbnail.name : ""
         await upload.updateData(req.body.mongoID, user._id, req.body.title, req.body.description, thumbnailName, user.name)
